feat(token-controls): relocate the shared color picker out of the Style Library
